feat: split code
- move converter functionality to converter.js file
- move validator functionality to validator.js file
